fix: Add ESLint disable comments for placeholder functions in TenantsPage
FIXED ESLINT ERRORS:

Added eslint-disable-next-line comments for placeholder functions with unused parameters:

1. viewTenantDetails(tenant)
   - Parameter will be used when detail view is implemented

2. editTenant(tenant)
   - Parameter will be used when edit functionality is implemented

3. deleteTenant(tenant)
   - Parameter will be used when delete functionality is implemented

REASON:
These are placeholder functions (marked with TODO) that will be implemented later.
The tenant parameter is necessary in the function signature for when they are implemented,
but currently triggers ESLint warnings since they're not yet used.

SOLUTION:
Added // eslint-disable-next-line no-unused-vars comments above each function
to suppress the warnings until the functions are fully implemented.

ESLINT STATUS:
✓ No linter errors

This is a temporary solution - when these functions are properly implemented,
the ESLint comments can be removed as the parameters will be actively used.